requirements.txt
pygame~=2.0.0
neat-python
matplotlib
numpy